requirements.txt
Django==4.2.28
openai==2.21.0
python-docx==1.2.0
pdfplumber==0.11.9
Pillow==12.1.1
gunicorn==25.1.0
whitenoise==6.12.0
python-dotenv==1.2.2
